PluggableIA NavMesh and Removed effects
started creating a AI (based on the "pluggable AI" tutorial from unity) that is a state machine.
For now there is only a patrolling state that makes the guard follow a path using the NavMeshAgent.
Also i removed the Post Processing, the glow effect, added a shader on guards (for now is a Asset, probably will do a custom one soon), then changed the wall shader to be used on mobile